0032 - Creating event-bus and send post requests
In this event-bus index.js file, we actually use axios to "send"
requests too. It's actually something we do from the front-end.
(axios is usually a package used in the front-end.) But we also create
a app.post() route inside event-bus index.js to handle the incoming
post requests to the event-bus too.