0032-sending a axios request from posts to event-bus
This will send a request from posts to the event-bus. We have a /events route on the event-bus to grab that request.
But as soon as the event-broker grab the request, the event broker will also send axios requests to other /envents routes
on other services. But we haven't created /events routes on any other services to grab the axios request sent by the event-brok
er