creating the comments moderation service
Create comment with pending status attached to it.

Emit “CommentCreated” event to the event bus.

Send that event to the moderation service and the query service.

Save it on the object on the query service with the status pending attached to it.

As soon as the commentCreated event is received by the moderation service, It should emit an event to the event bus called “CommentModerated” with the updated status attached to it based on the logic we put inside the post(’/events’) route on the moderation service. In this case, the logic is there for filtering out bad words.

Now we should pass the CommentModerated event to the event bus and it will pass it to the Comments service

In the post(‘/events’) route on the comment service, we should write the logic to filter out the actual comment that matches the id sent by CommentModerated event and update the status of that comment with the status inside that CommentModerated event.

Now we need to send a CommentUpdated event from the comments service to the event bus with the commentId, status, postId and content attached to it

Now that event should be received by the query service. Then we need to write a logic inside that query service to filter out the comment corresponding to the id on the CommentUpdated event and update the entire comment (both content and the status) from the newly updated data. Cause CommentUpdated event can be happen not only from filtering out bad words, but also from other kinds of things like upvoted, reported, liked and all sorts of things. So we should generalize this update to correspond to any kind of comment update so that we need to update each and every aspect of the comment from the new comment sent by the CommentUpdated event.